+++ App/Screens/Home/SmallButton/SmallButton.tsx
@@ -23,7 +23,7 @@ import {
   TouchableOpacityProps
 } from 'react-native';
 
-import * as theme from '../../../utils/theme';
+import * as theme from '../../../util/theme';
 
 interface SmallButtonProps extends TouchableOpacityProps {
   icon: string;
