+++ App/stores/fetchApi/index.ts
@@ -14,26 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with Sh**t! I Smoke.  If not, see <http://www.gnu.org/licenses/>.
 
-import React, { createContext, useState } from 'react';
-
-import { noop } from '../util/noop';
-
-interface Context {
-  error?: string;
-  setError: (error?: string) => void;
-}
-
-export const ErrorContext = createContext<Context>({
-  error: undefined,
-  setError: noop
-});
-
-export function ErrorContextProvider ({ children }: { children: JSX.Element }) {
-  const [error, setError] = useState<string | undefined>(undefined);
-
-  return (
-    <ErrorContext.Provider value={{ error, setError }}>
-      {children}
-    </ErrorContext.Provider>
-  );
-}
+export * from './fetchApi';
